+++ index.ts
@@ -1,7 +1,7 @@
 import inputData from "./src/input.json"
 import expectedResult from "./src/output.json"
 import { evaluateSolution } from "./src/evaluate"
-import solution from "./src/solutions/template"
+import solution from "./src/solutions/TortillaGuy"
 
 const solver = evaluateSolution(inputData as any, expectedResult as any)
 const score = solver(solution)
